# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My POS App API",
        "description": "API documentation for My POS App backend.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/purchases": {
            "get": {
                "tags": [
                    "Purchase"
                ],
                "summary": "List all purchases",
                "operationId": "4af8610ecc39e89061fd641f3584cca1",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Purchase"
                ],
                "summary": "Create a new purchase",
                "operationId": "40a50a92dc96d3d902a69d4c255c0e47",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Purchase created"
                    }
                }
            }
        },
        "/purchases/{id}": {
            "get": {
                "tags": [
                    "Purchase"
                ],
                "summary": "Get a single purchase",
                "operationId": "bd83db79c7145fd9aaf1f7a57d7955d6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Purchase not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Purchase"
                ],
                "summary": "Update a purchase",
                "operationId": "df22c739fc38177f931523f61fbb5ad8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Purchase updated"
                    },
                    "404": {
                        "description": "Purchase not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Purchase"
                ],
                "summary": "Delete a purchase",
                "operationId": "f2f073aaf73aef3f7e9c03dc3f954b07",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Purchase deleted"
                    },
                    "404": {
                        "description": "Purchase not found"
                    }
                }
            }
        },
        "/purchase-details": {
            "get": {
                "tags": [
                    "PurchaseDetail"
                ],
                "summary": "List all purchase details",
                "operationId": "e4260a5ca928b7746378e77a635fb08e",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "PurchaseDetail"
                ],
                "summary": "Create a new purchase detail",
                "operationId": "41027ecabc97988bf563d1463989ea89",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Purchase detail created"
                    }
                }
            }
        },
        "/purchase-details/{id}": {
            "get": {
                "tags": [
                    "PurchaseDetail"
                ],
                "summary": "Get a single purchase detail",
                "operationId": "dcae3473a624b05ff3e92e6d6ad43c99",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Purchase detail not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "PurchaseDetail"
                ],
                "summary": "Update a purchase detail",
                "operationId": "afd59704d3de1e93d628e834f4075d34",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Purchase detail updated"
                    },
                    "404": {
                        "description": "Purchase detail not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "PurchaseDetail"
                ],
                "summary": "Delete a purchase detail",
                "operationId": "2e751ed5e692f4271d88901c04bff449",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Purchase detail deleted"
                    },
                    "404": {
                        "description": "Purchase detail not found"
                    }
                }
            }
        },
        "/purchase-invoices": {
            "get": {
                "tags": [
                    "PurchaseInvoice"
                ],
                "summary": "List all purchase invoices",
                "operationId": "b5a0047760cf3010b4028f7de0091958",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "PurchaseInvoice"
                ],
                "summary": "Create a new purchase invoice",
                "operationId": "c378b2a7f204417100a6aa7b6093a705",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Purchase invoice created"
                    }
                }
            }
        },
        "/purchase-invoices/{id}": {
            "get": {
                "tags": [
                    "PurchaseInvoice"
                ],
                "summary": "Get a single purchase invoice",
                "operationId": "2fc91ceb445d627ff61d6268c7923ace",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Purchase invoice not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "PurchaseInvoice"
                ],
                "summary": "Update a purchase invoice",
                "operationId": "d7599806141266e0417e3d4bbf74b44e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Purchase invoice updated"
                    },
                    "404": {
                        "description": "Purchase invoice not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "PurchaseInvoice"
                ],
                "summary": "Delete a purchase invoice",
                "operationId": "0f2c880419a01a5429e44141c77dae8f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Purchase invoice deleted"
                    },
                    "404": {
                        "description": "Purchase invoice not found"
                    }
                }
            }
        },
...
    },
...
}